FAD-5976 fixed handling 4xx error, addressed test feedback
