+++ src/pages/auth/components/tests/Tfaform.test.js
@@ -5,7 +5,8 @@ import { TfaForm } from '../TfaForm';
 
 const props = {
   tfaPending: false,
-  handleSubmit: jest.fn()
+  handleSubmit: jest.fn(),
+  error: false
 };
 
 let wrapper;
@@ -26,4 +27,9 @@ it('renders correctly verifying tfa', () => {
 it('calls correct method on submit', () => {
   wrapper.find('form').first().simulate('submit');
   expect(props.handleSubmit).toHaveBeenCalledTimes(1);
+});
+
+it('should display error when form submission fails', () => {
+  wrapper.setProps({ error: { message: 'tfa failed' }});
+  expect(wrapper).toMatchSnapshot();
 });
